+++ src/Controller/StaticController.php
@@ -4,10 +4,6 @@ declare(strict_types=1);
 
 namespace App\Controller;
 
-use App\Entity\User;
-use App\Enum\RolesEnum;
-use App\Repository\UserEntityRepository;
-use Doctrine\ORM\EntityManagerInterface;
 use League\CommonMark\Environment\Environment;
 use League\CommonMark\Extension\CommonMark\CommonMarkCoreExtension;
 use League\CommonMark\MarkdownConverter;
@@ -98,84 +94,6 @@ class StaticController extends AbstractController
         return $this->render('static/unfold.html.twig');
     }
 
-    #[Route('/essayist', name: 'app_static_essayist', methods: ['GET'])]
-    public function essayist(Request $request, UserEntityRepository $userRepository): Response
-    {
-        $user = $this->getUser();
-        $roles = $user instanceof User ? $user->getRoles() : [];
-
-        $launchDate = new \DateTimeImmutable('2026-06-01');
-
-        return $this->render('static/essayist.html.twig', [
-            'isMember'         => in_array(RolesEnum::ESSAYIST_MEMBER->value, $roles, true),
-            'isPending'        => in_array(RolesEnum::ESSAYIST_CANDIDATE->value, $roles, true),
-            'isEarlyBird'      => in_array(RolesEnum::ESSAYIST_EARLY_BIRD->value, $roles, true),
-            'memberCount'      => $userRepository->countByRole(RolesEnum::ESSAYIST_MEMBER->value),
-            'joinStatus'       => $request->query->get('join_status'),
-            'launchDate'       => $launchDate,
-            'isLaunched'       => new \DateTimeImmutable() >= $launchDate,
-            'earlyBirdDeadline'=> new \DateTimeImmutable('2026-05-31'),
-        ]);
-    }
-
-    #[Route('/essayist/early-bird', name: 'app_static_essayist_early_bird', methods: ['POST'])]
-    public function claimEarlyBird(
-        Request $request,
-        EntityManagerInterface $entityManager,
-    ): Response {
-        $user = $this->getUser();
-        if (!$user instanceof User) {
-            return $this->redirectToRoute('app_static_essayist', ['join_status' => 'login_required']);
-        }
-
-        $token = (string) $request->request->get('_token', '');
-        if (!$this->isCsrfTokenValid('essayist_early_bird', $token)) {
-            return $this->redirectToRoute('app_static_essayist', ['join_status' => 'invalid_csrf']);
-        }
-
-        $roles = $user->getRoles();
-        if (in_array(RolesEnum::ESSAYIST_EARLY_BIRD->value, $roles, true)) {
-            return $this->redirectToRoute('app_static_essayist', ['join_status' => 'already_early_bird']);
-        }
-
-        $user->addRole(RolesEnum::ESSAYIST_EARLY_BIRD->value);
-        $user->addRole(RolesEnum::ESSAYIST_MEMBER->value);
-        $entityManager->persist($user);
-        $entityManager->flush();
-
-        return $this->redirectToRoute('app_static_essayist', ['join_status' => 'early_bird_claimed']);
-    }
-
-    #[Route('/essayist/request-access', name: 'app_static_essayist_request_access', methods: ['POST'])]
-    public function requestEssayistAccess(
-        Request $request,
-        EntityManagerInterface $entityManager,
-    ): Response {
-        $user = $this->getUser();
-        if (!$user instanceof User) {
-            return $this->redirectToRoute('app_static_essayist', ['join_status' => 'login_required']);
-        }
-
-        $token = (string) $request->request->get('_token', '');
-        if (!$this->isCsrfTokenValid('essayist_request_access', $token)) {
-            return $this->redirectToRoute('app_static_essayist', ['join_status' => 'invalid_csrf']);
-        }
-
-        $roles = $user->getRoles();
-        if (in_array(RolesEnum::ESSAYIST_MEMBER->value, $roles, true)) {
-            return $this->redirectToRoute('app_static_essayist', ['join_status' => 'already_member']);
-        }
-
-        if (in_array(RolesEnum::ESSAYIST_CANDIDATE->value, $roles, true)) {
-            return $this->redirectToRoute('app_static_essayist', ['join_status' => 'already_pending']);
-        }
-
-        $user->addRole(RolesEnum::ESSAYIST_CANDIDATE->value);
-        $entityManager->persist($user);
-        $entityManager->flush();
-
-        return $this->redirectToRoute('app_static_essayist', ['join_status' => 'request_created']);
-    }
 
     #[Route('/start-blog', name: 'start_blog')]
     public function startBlog(): Response